stable release

+++ src/app/components/DeleteDialouge.tsx
@@ -1,9 +1,10 @@
 import { Fragment, useState } from "react";
+import { useSession } from 'next-auth/react';
 import { Dialog, Transition } from "@headlessui/react";
 import { doc, deleteDoc } from 'firebase/firestore';
 import { ref, deleteObject } from 'firebase/storage';
-import { db, storage } from "../firebase";
-import { CgSpinner } from 'react-icons/cg';
+import { db, storage } from "../lib/firebase";
+import StylistButton from "./MEXTUI/StylistButton";
 import { MdDelete } from 'react-icons/md';
 import { RxCross2 } from 'react-icons/rx';
 
@@ -11,8 +12,9 @@ interface DeleteDialougeProps {
   deleteDialouge: {
     isOpen: boolean;
     siteID: string;
+    imageUrl: string;
   }
-  setDeleteDialouge: (e: any) => void;
+  setDeleteDialouge: (newState: { isOpen: boolean, siteID: string, imageUrl: string }) => void;
   getSiteList: () => void;
 }
 
@@ -29,22 +31,29 @@ const DeleteDialouge = ({ deleteDialouge, setDeleteDialouge, getSiteList }: Dele
   });
   const [isLoading, setLoading] = useState<boolean>(false);
 
+  const { data: session } = useSession();
+
   const handleDelete = async() => {
+    if (session?.user?.name === "Guest") return;
     setLoading(true);
     try {
       const docRef = doc(db, 'siteinfo', deleteDialouge.siteID);
       await deleteDoc(docRef);
-      const storageRef = ref(storage, `images/${deleteDialouge.siteID}`);
-      await deleteObject(storageRef);
+      if (deleteDialouge.imageUrl) {
+        const storageRef = ref(storage, `images/${deleteDialouge.siteID}`);
+        await deleteObject(storageRef);
+      }
       setStatus({ type: true, value: "Deleted successfully !"});
     } catch (error: any) {
       setStatus({ type: false, value: error.toString()});
+      setLoading(false);
       return;
     }
     setLoading(false);
     setTimeout(() => {
       getSiteList();
-      setDeleteDialouge({ isOpen: false, siteID: "" });
+      setStatus({ type: true, value: ""});
+      setDeleteDialouge({ isOpen: false, siteID: "", imageUrl: '' });
     }, 1000);
 
   }
@@ -78,23 +87,17 @@ const DeleteDialouge = ({ deleteDialouge, setDeleteDialouge, getSiteList }: Dele
               <Dialog.Panel className="mx-auto max-w-2xl container rounded bg-white p-6">
               <div className="flex justify-between items-center mb-5 text-center">
                 <Dialog.Title className="text-2xl font-semibold text-center w-[92%]">Delete confirmation</Dialog.Title>
-                <button type="button" onClick={() => setDeleteDialouge({ isOpen: false, siteID: ""})} className="w-[8%]">
-                  <RxCross2 size="lg" className="h-9 w-9 text-gray-500 hover:text-black transition-all" />
+                <button type="button" onClick={() => setDeleteDialouge({ isOpen: false, siteID: "", imageUrl: ''})} className="w-[8%]">
+                  <RxCross2 size={32} className="text-gray-500 hover:text-black transition-all" />
                 </button>
               </div>
-
-              <p>Are you sure to Delete this site from switch panel? This can't be undone!</p>
-              <p>All data related to this site will be deleted.</p>
-
-
+              <p>Are you sure to Delete this site from switch panel? This can't be undone!<br /> All data related to this site will be deleted.</p>
               <div className="flex justify-between items-center col-span-1 lg:col-span-2 mt-5">
                 <p className={status.type ? "text-emerald-700" : "text-red-500"}>{status.value}</p>
-                <button type="button" disabled={isLoading} onClick={handleDelete} className="px-2 py-1 text-white rounded flex items-center space-x-1 bg-red-500 disabled:bg-gray-500 hover:bg-red-700 transition-all">
-                  {isLoading ? <CgSpinner size="lg" className="h-4 w-4 animate-spin" /> : <MdDelete size="lg" className="h-4 w-4" />}
-                  <p>{isLoading ? "Deleting..." : "Delete"}</p>
-                </button>
+                <StylistButton onClick={handleDelete} disabled={isLoading} loading={isLoading} label="Confirm Delete" loadingLabel="Deleting..." size="sm" bgColor="#d60927" space={3} bgColorOnHover="#a3051d" childrenBeforeLabel>
+                  <MdDelete size={16} />
+                </StylistButton>
               </div>
-
               </Dialog.Panel>
             </Transition.Child>
           </div>
